Stop details from injecting tags into the HTML
We had a service detail that included a message about an HTML tag
(<title>). That caused the rest of the nagdash page to not display. The
fix is to escape the HTML entities in the detail text.

I'm no PHP expert (or even novice), so I'm not certain about how to
write what I did, but it seems pretty straightforward. The function used
is from https://stackoverflow.com/questions/10551116/html-php-escape-and-symbols-while-echoing